Clean up EventController.kt added Event.kt Added Doc
