+++ src/commonMain/kotlin/GameModule.kt
@@ -7,7 +7,7 @@ import kotlin.reflect.*
 
 object GameModule : Module() {
 
-    override val mainScene: KClass<out Scene> = Physics_Sample::class
+    override val mainScene: KClass<out Scene> = TestScene::class
     override val size: SizeInt = SizeInt(1920,1080)
     //override val bgcolor: RGBA = Colors.WHITE
