+++ src/commonMain/kotlin/Scenes/Test_Scene.kt
@@ -3,11 +3,11 @@ package Scenes
 //import fsm.Entity
 import com.soywiz.korge.scene.Scene
 import com.soywiz.korge.view.Container
+import eventBus.*
+import kotlinx.coroutines.*
 
 class TestScene : Scene() {
     override suspend fun Container.sceneInit() {
-
-        //val bus = EventBus(this@TestScene)
 
         /*val testPlayer = Player.build("Characters/Test.xml", this@TestScene).apply { xy(100, 100); scale = 0.1 }
         val testEnemy = Enemy.build("Characters/Test.xml", this@TestScene).apply { xy(300, 300); scale = 0.1 }
