+++ src/commonMain/kotlin/actor/actors/Enemy.kt
@@ -112,15 +112,7 @@ class Enemy(
     }
 
     override fun initEvents() {
-        bus.register<PlayerCollision> { onPlayerCollision(it.aabb) }
-        bus.register<EnemyCollision> { onPlayerCollision(it.aabb) }
-        bus.register<GroundCollision> { onGroundCollision() }
-        bus.register<PlatformCollision> { onPlatformCollision(it.aabb) }
-        bus.register<NormalAttackCollision> { onNormalAttackCollision(it.damage) }
-        bus.register<RangedAttackCollision> { onRangedAttackCollision(it.damage) }
-        bus.register<SpecialAttackCollision> { onSpecialAttackCollision(it.damage) }
 
-        bus.register<StateTransition> { doStateChange(it.state) }
     }
 
     //maybe?
@@ -254,7 +246,7 @@ class Enemy(
     override fun executeState_turn(dt: Double) {
         timer += 1
        // physics.update(dt)
-        if (model.animation.isCompleted) bus.send(StateTransition(walkState))
+
     }
 
     override fun endState_turn() { /* Nothing in here */
@@ -308,7 +300,7 @@ class Enemy(
         //check if he collides with something -> this can take damage
         //physics.update(dt)
         if (model.animation.isCompleted) {
-            bus.send(StateTransition(idleState))
+
         }
     }
 
@@ -326,7 +318,7 @@ class Enemy(
         timer += 1
         //physics.update(dt)
         if (model.animation.isCompleted) {
-            bus.send(StateTransition(idleState))
+
         }
     }
 
@@ -365,7 +357,7 @@ class Enemy(
         timer += 1
         //physics.update(dt)
         if (model.animation.isCompleted) {
-            bus.send(StateTransition(idleState))
+
         }
     }
