+++ src/commonMain/kotlin/eventBus/Event.kt
@@ -1,0 +1,3 @@
+package eventBus
+
+abstract class Event {}
